+++ src/main/java/net/spell_engine/client/SpellEngineClient.java
@@ -7,7 +7,10 @@ import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLifecycleEvents;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.screen.ingame.HandledScreens;
 import net.minecraft.client.render.block.entity.BlockEntityRendererFactories;
+import net.minecraft.item.BowItem;
+import net.minecraft.item.CrossbowItem;
 import net.minecraft.item.Items;
+import net.minecraft.registry.Registries;
 import net.spell_engine.SpellEngineMod;
 import net.spell_engine.client.animation.AnimationRegistry;
 import net.spell_engine.client.render.ModelPredicateHelper;
@@ -41,15 +44,21 @@ public class SpellEngineClient {
         ClientLifecycleEvents.CLIENT_STARTED.register((client) -> {
             var resourceManager = MinecraftClient.getInstance().getResourceManager();
             AnimationRegistry.load(resourceManager);
+            injectRangedWeaponModelPredicates();
         });
 
         HandledScreens.register(SpellBindingScreenHandler.HANDLER_TYPE, SpellBindingScreen::new);
         BlockEntityRendererFactories.register(SpellBindingBlockEntity.ENTITY_TYPE, SpellBindingBlockEntityRenderer::new);
     }
 
-    private static void injectVanillaModelPredicates() {
-        // Override vanilla model predicates to inject spell casting progress
-        ModelPredicateHelper.injectBowSkillUsePredicate(Items.BOW);
-        ModelPredicateHelper.injectCrossBowSkillUsePredicate(Items.CROSSBOW);
+    private static void injectRangedWeaponModelPredicates() {
+        for(var itemId: Registries.ITEM.getIds()) {
+            var item = Registries.ITEM.get(itemId);
+            if (item instanceof BowItem) {
+                ModelPredicateHelper.injectBowSkillUsePredicate(item);
+            } else if (item instanceof CrossbowItem) {
+                ModelPredicateHelper.injectCrossBowSkillUsePredicate(item);
+            }
+        }
     }
 }